fix 🐛 correção para a criação das migração de forma correta

+++ routes/api.php
@@ -22,6 +22,7 @@ use App\Http\Controllers\TecnologiasController;
 use App\Http\Controllers\EntregaClienteController;
 use App\Http\Controllers\ProdutoUsuarioController;
 use App\Http\Controllers\TipoEntregaUserController;
+
 /**
  * Rotas da API do Sistema de Gerenciamento de Créditos
  *
@@ -76,12 +77,13 @@ Route::middleware(['auth.jwt'])->group(function () {
      * @middleware permission
      */
     Route::prefix('vendas_creditos')->group(function () {
-        Route::get('/', [CreditSaleController::class, 'index']);
-        Route::get('/cliente/{id}', [CreditSaleController::class, 'buscarTransacoesPorCliente']);
-        Route::post('/', [CreditSaleController::class, 'store']);
+        Route::get('/', [CreditSaleController::class, 'index'])->middleware('role:admin,producao');
+        Route::get('/cliente/{id}', [CreditSaleController::class, 'buscarTransacoesPorCliente'])->middleware('role:admin,producao');
+        Route::post('/', [CreditSaleController::class, 'store'])->middleware('role:admin,producao');
+        Route::post('{id}/cancelar', [CreditSaleController::class, 'cancelVendaCreditos'])->middleware('role:admin,producao');
         Route::get('/{creditSale}', [CreditSaleController::class, 'show']);
-        Route::put('/{creditSale}', [CreditSaleController::class, 'update']);
-        Route::delete('/{creditSale}', [CreditSaleController::class, 'destroy']);
+        Route::put('/{creditSale}', [CreditSaleController::class, 'update'])->middleware('role:admin,producao');
+        Route::delete('/{creditSale}', [CreditSaleController::class, 'destroy'])->middleware('role:admin,producao');
     });
 
 
@@ -94,35 +96,31 @@ Route::middleware(['auth.jwt'])->group(function () {
      * @middleware permission
      */
     Route::prefix('remessas')->group(function () {
-            // 👤 Cliente: Cria remessa e consulta status
+        // 👤 Cliente: Cria remessa e consulta status
         Route::get('/', [RemessaController::class, 'index'])->middleware('role:cliente,admin,producao');
         Route::post('/', [RemessaController::class, 'store'])->middleware('role:cliente,admin');
 
         // Usuários com acesso para produzir
         Route::get('/tarefas-disponiveis', [RemessaController::class, 'tarefasDisponiveis'])->middleware('role:admin,producao');
-        Route::get('/minhas-tarefas', [RemessaController::class, 'minhasTarefas']);
-        Route::get('/tarefas-expedicoes', [RemessaController::class, 'tarefasEmExpedicao'])->middleware('role:admin,producao');
+        Route::get('/minhas-tarefas', [RemessaController::class, 'minhasTarefas'])->middleware('role:admin,producao');
+        Route::get('/tarefas-expedicoes', [RemessaController::class, 'tarefasEmExpedicao'])->middleware('role:admin,expedicao');
+        Route::get('/tarefas-balcao', [RemessaController::class, 'tarefasBalcao'])->middleware('role:admin,recepcao');
+
         // 📷 Uploads (liberado para admin e produção se necessário)
         Route::post('/{remessa}/upload-fotos', [RemessaFotoController::class, 'store'])->middleware('role:admin,producao');
         Route::post('/{remessa}/upload-planilha', [RemessaPlanilhaController::class, 'store'])->middleware('role:admin,producao');
 
 
         // 📄 Detalhes da remessa
-        Route::get('/{remessa}', [RemessaController::class, 'show'])->middleware('role:admin,cliente');
+        Route::get('/{remessa}', [RemessaController::class, 'show'])->middleware('role:admin,producao,cliente');
 
         // ✏️ Atualização e exclusão (restrito a admin)
         Route::put('/{id}', [RemessaController::class, 'update'])->middleware('role:admin,producao');
         Route::delete('/{remessa}', [RemessaController::class, 'destroy'])->middleware('role:admin');
-
-        // // 🚦 Transições de status protegidas por role específica
-        // Route::post('/{remessa}/produzir', [RemessaStatusController::class, 'produzir'])->middleware('role:producao');
-        // Route::post('/{remessa}/expedir', [RemessaStatusController::class, 'expedir'])->middleware('role:producao');
-        // Route::post('/{remessa}/liberar', [RemessaStatusController::class, 'liberar'])->middleware('role:expedicao');
-        // Route::post('/{remessa}/finalizar', [RemessaStatusController::class, 'finalizar'])->middleware('role:recepcao');
     });
 
 
-      /**
+    /**
      * Rotas de Modelos
      * Gerenciamento de modelos
      *
@@ -131,8 +129,8 @@ Route::middleware(['auth.jwt'])->group(function () {
      */
     Route::prefix('modelo-tecnico')->group(function () {
         Route::get('/', [ModeloTecnicosController::class, 'index'])->middleware('role:admin,cliente,producao');
-        Route::get('/clientes/{id}', [ModeloTecnicosController::class, 'buscarPorUsuarios'])->middleware('role:admin,cliente,producao');
-        Route::get('/cliente/unico/{id}', [ModeloTecnicosController::class, 'buscarModeloUnico'])->middleware('role:admin,cliente,producao');
+        Route::get('/clientes/{id}', [ModeloTecnicosController::class, 'buscarPorUsuarios'])->middleware('role:admin,cliente,producao,expedicao');
+        Route::get('/cliente/unico/{id}', [ModeloTecnicosController::class, 'buscarModeloUnico'])->middleware('role:admin,cliente,producao,expedicao');
 
         Route::post('/', [ModeloTecnicosController::class, 'store'])->middleware('role:admin,cliente,producao');
         Route::get('/{id}', [ModeloTecnicosController::class, 'show']);
@@ -151,6 +149,7 @@ Route::middleware(['auth.jwt'])->group(function () {
     Route::prefix('modelos-tecnicos-campos-variaveis')->group(function () {
         Route::get('/', [ModelosTecnicosCamposVariaveisController::class, 'index']);
         Route::post('/', [ModelosTecnicosCamposVariaveisController::class, 'store'])->middleware('role:admin,cliente,producao');
+        Route::post('/reorganizar/{modelo_tecnico_id}', [ModelosTecnicosCamposVariaveisController::class, 'reorganizarOrdem']);
         Route::get('/{id}', [ModelosTecnicosCamposVariaveisController::class, 'show']);
         Route::put('/{id}', [ModelosTecnicosCamposVariaveisController::class, 'update'])->middleware('role:admin,cliente,producao');
         Route::delete('/{id}', [ModelosTecnicosCamposVariaveisController::class, 'destroy'])->middleware('role:admin,cliente,producao');
@@ -214,7 +213,7 @@ Route::middleware(['auth.jwt'])->group(function () {
     });
 
 
-     /**
+    /**
      * Rotas de tipos de entregas
      * Gerenciamento de tipos de entregas (CRUD)
      *
@@ -232,7 +231,7 @@ Route::middleware(['auth.jwt'])->group(function () {
         Route::delete('/{id}', [TipoEntregaController::class, 'destroy']);
     });
 
-     /**
+    /**
      * Rotas de Endereços
      * Gerenciamento de endereços (CRUD)
      *
@@ -249,7 +248,7 @@ Route::middleware(['auth.jwt'])->group(function () {
     });
 
 
-     /**
+    /**
      * Rotas de Endereços
      * Gerenciamento de endereços (CRUD)
      *
@@ -264,7 +263,7 @@ Route::middleware(['auth.jwt'])->group(function () {
         Route::delete('/{id}', [TecnologiasController::class, 'destroy']);
     });
 
-        /**
+    /**
      * Rotas de vincular produtos ao cliente
      * Gerenciamento de vincular produtos ao cliente (CRUD)
      *
@@ -279,7 +278,7 @@ Route::middleware(['auth.jwt'])->group(function () {
 
 
 
-         /**
+    /**
      * Rotas de entreas
      * Gerenciamento de entregas de remessas (CRUD)
      *
@@ -293,7 +292,6 @@ Route::middleware(['auth.jwt'])->group(function () {
         Route::put('/{id}', [EntregaClienteController::class, 'update']);
         Route::delete('/{id}', [EntregaClienteController::class, 'destroy']);
     });
-
 });
